feat: #72 문서 공유 기능 utils로 분리
