+++ frontend/src/components/editorHeader/EditorHeader.tsx
@@ -2,19 +2,10 @@ import React from 'react';
 import styled from 'styled-components';
 import { SiteLogo } from '../siteLogo';
 import useTitle from '../../hooks/useTitle';
+import { copyURLPath } from '../../utils/utils';
 
 const EditorHeader = () => {
   const { title, onTitleChange } = useTitle('Untitled');
-  const copyURLPath = () => {
-    try {
-      const url = window.location.href;
-      navigator.clipboard
-        .writeText(url)
-        .then(() => alert('주소가 복사되었습니다.\n문서를 다른 사람에게 공유해보세요!'));
-    } catch (e) {
-      alert('주소 복사에 실패했어요 ㅠㅠ');
-    }
-  };
 
   return (
     <>
